made the BUCKET configurable for the common crawl example

+++ scripts/00_package.sh
@@ -1,6 +1,6 @@
 #!/usr/bin/env bash
 
-BUCKET=s3://offline-elastic-world-index
+[[ -z "$BUCKET" ]] && { echo "BUCKET is empty exiting" ; exit 1; }
 
 rm -Rf offline_worker* && rm -Rf data/repo/* && sbt clean assembly
 aws s3 cp offline-index-cc/target/scala-2.11/offline-index-cc-0.1-SNAPSHOT.jar ${BUCKET}/deploy/
